requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.3
opencv-python==4.4.0.44
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
Shapely==1.7.1
six==1.15.0
tifffile==2020.10.1
